Update IntroR.R
There were a couple typos, and since I was making a PR anyway, I added a bunch of commas to the non-essential clauses. Line 183: “# Take a look at the History list, on the tab to the left.” —> Is it left for you? It’s right for me. I changed it to right just in case but obv please double check that one on your system.